refactor(app): move chrome widget source-of-truth into widgets/ (#3594)
Reverse the inverted dependency direction for the three shared chrome
widgets CtDialogShell, CtPanel and CtNinePatchButton: their
implementations now live in app/lib/widgets/, and the
features/game/widgets/chrome/ files become thin re-export shims pointing
back up to widgets/ (features -> widgets).

- Drop the three former bridge entries from the
  check_app_widget_imports allowlist (now real implementations that
  import only config/ + widgets/).
- Add check_chrome_source_of_truth_test.dart to pin the reversed
  direction and prevent regression.

Refs #3594
